refactor : LineStations Embedded 설정
