+++ src/main/java/nextstep/subway/domain/line/Line.java
@@ -1,6 +1,7 @@
 package nextstep.subway.domain.line;
 
 import nextstep.subway.domain.LineStation.LineStation;
+import nextstep.subway.domain.LineStation.LineStations;
 import nextstep.subway.domain.common.BaseEntity;
 import nextstep.subway.domain.common.Distance;
 import nextstep.subway.domain.section.Section;
@@ -8,8 +9,6 @@ import nextstep.subway.domain.section.Sections;
 import nextstep.subway.domain.station.Station;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 public class Line extends BaseEntity {
@@ -29,16 +28,11 @@ public class Line extends BaseEntity {
     @JoinColumn(name = "down_station_id")
     private Station downStation;
 
-    @OneToMany(
-            mappedBy = "line",
-            fetch = FetchType.LAZY,
-            cascade = CascadeType.ALL,
-            orphanRemoval = true
-    )
-    private List<LineStation> lineStations = new ArrayList<>();
+    @Embedded
+    private LineStations lineStations = LineStations.create();
 
     @Embedded
-    private Sections sections = Sections.empty();
+    private Sections sections = Sections.create();
 
     @Embedded
     private Distance distance;
@@ -86,26 +80,11 @@ public class Line extends BaseEntity {
     }
 
     public void setTerminus(Station upStation, Station downStation) {
-        validateDuplicateStation(upStation);
-        validateDuplicateStation(downStation);
-
         lineStations.add(LineStation.create(this, upStation));
         lineStations.add(LineStation.create(this, downStation));
 
         this.upStation = upStation;
         this.downStation = downStation;
-    }
-
-    private void validateDuplicateStation(Station station) {
-        if (isExistSameStation(station)) {
-            throw new IllegalArgumentException("이미 노선에 포함된 지하철역입니다. station : " + station);
-        }
-    }
-
-    private boolean isExistSameStation(Station station) {
-        return lineStations.stream()
-                .map(LineStation::getStation)
-                .anyMatch(storedStation -> storedStation.equals(station));
     }
 
     public void modify(String name, String color) {
